refactor(INJI-434): extracts telemetry constants to a seperate file
