+++ machines/QrLoginMachine.ts
@@ -23,10 +23,10 @@ import {
 import i18n from '../i18n';
 import {parseMetadatas, VCMetadata} from '../shared/VCMetadata';
 import {
-  TelemetryConstants,
   getEndEventData,
   sendEndEvent,
 } from '../shared/telemetry/TelemetryUtils';
+import {TelemetryConstants} from '../shared/telemetry/TelemetryConstants';
 
 const model = createModel(
   {
