+++ machines/VCItemMachine/EsignetMosipVCItem/EsignetMosipVCItemMachine.ts
@@ -30,8 +30,9 @@ import {
   sendEndEvent,
   sendInteractEvent,
   sendStartEvent,
-  TelemetryConstants,
 } from '../../../shared/telemetry/TelemetryUtils';
+import {TelemetryConstants} from '../../../shared/telemetry/TelemetryConstants';
+
 import {API_URLS} from '../../../shared/api';
 
 const model = createModel(
