+++ machines/VCItemMachine/ExistingMosipVCItem/ExistingMosipVCItemMachine.ts
@@ -34,11 +34,12 @@ import {
   getEndEventData,
   getStartEventData,
   sendEndEvent,
-  TelemetryConstants,
   sendInteractEvent,
   getInteractEventData,
   sendStartEvent,
 } from '../../../shared/telemetry/TelemetryUtils';
+import {TelemetryConstants} from '../../../shared/telemetry/TelemetryConstants';
+
 import {API_URLS} from '../../../shared/api';
 
 const model = createModel(
