+++ screens/Issuers/IssuersScreen.tsx
@@ -17,12 +17,12 @@ import {
   Protocols,
 } from '../../shared/openId4VCI/Utils';
 import {
-  TelemetryConstants,
   getInteractEventData,
   getStartEventData,
   sendInteractEvent,
   sendStartEvent,
 } from '../../shared/telemetry/TelemetryUtils';
+import {TelemetryConstants} from '../../shared/telemetry/TelemetryConstants';
 
 export const IssuersScreen: React.FC<
   HomeRouteProps | RootRouteProps
