+++ screens/WelcomeScreenController.ts
@@ -18,8 +18,8 @@ import {
   getInteractEventData,
   sendInteractEvent,
   sendStartEvent,
-  TelemetryConstants,
 } from '../shared/telemetry/TelemetryUtils';
+import {TelemetryConstants} from '../shared/telemetry/TelemetryConstants';
 
 export function useWelcomeScreen(props: RootRouteProps) {
   const {appService} = useContext(GlobalContext);
